# api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"_36_Enums.QuestionType": {
				"type": "string",
				"enum": [
					"SC",
					"MC",
					"TEXT"
				]
			},
			"QuestionType": {
				"$ref": "#/components/schemas/_36_Enums.QuestionType"
			},
			"_36_Enums.Subject": {
				"type": "string",
				"enum": [
					"ENGLISH",
					"IT"
				]
			},
			"Subject": {
				"$ref": "#/components/schemas/_36_Enums.Subject"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Paste the JWT access token obtained from /auth/login"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"description": "A description of QUIZZZ API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/questions": {
			"get": {
				"operationId": "ListQuestions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"total": {
											"type": "number",
											"format": "double"
										},
										"items": {
											"items": {
												"properties": {
													"explanation": {
														"type": "string",
														"nullable": true
													},
													"text": {
														"type": "string"
													},
													"id": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"explanation",
													"text",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										}
									},
									"required": [
										"total",
										"items"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Question"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "examId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"explanation": {
											"type": "string",
											"nullable": true
										},
										"text": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"explanation",
										"text",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Question"
				],
				"security": [
					{
						"bearerAuth": [
							"TEACHER",
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"points": {
										"type": "number",
										"format": "double"
									},
									"choices": {
										"items": {
											"properties": {
												"order": {
													"type": "number",
													"format": "double"
												},
												"isCorrect": {
													"type": "boolean"
												},
												"content": {
													"type": "string"
												}
											},
											"required": [
												"isCorrect",
												"content"
											],
											"type": "object"
										},
										"type": "array"
									},
									"subject": {
										"$ref": "#/components/schemas/Subject"
									},
									"type": {
										"$ref": "#/components/schemas/QuestionType"
									},
									"explanation": {
										"type": "string",
										"nullable": true
									},
									"text": {
										"type": "string"
									},
									"examId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"choices",
									"text",
									"examId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/questions/{id}": {
			"get": {
				"operationId": "GetQuestionById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"choices": {
											"items": {
												"properties": {
													"order": {
														"type": "number",
														"format": "double"
													},
													"isCorrect": {
														"type": "boolean"
													},
													"content": {
														"type": "string"
													},
													"id": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"order",
													"isCorrect",
													"content",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										},
										"explanation": {
											"type": "string",
											"nullable": true
										},
										"text": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"choices",
										"explanation",
										"text",
										"id"
									],
									"type": "object",
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Question not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Question"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"explanation": {
											"type": "string",
											"nullable": true
										},
										"text": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"explanation",
										"text",
										"id"
									],
									"type": "object",
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Question not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Question"
				],
				"security": [
					{
						"bearerAuth": [
							"TEACHER",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"type": {
										"$ref": "#/components/schemas/QuestionType"
									},
									"explanation": {
										"type": "string",
										"nullable": true
									},
									"text": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"message"
									],
									"type": "object",
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Question not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Question"
				],
				"security": [
					{
						"bearerAuth": [
							"TEACHER",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}